src/requirements.txt
pyodbc==5.1.0
pandas==2.2.2
sqlalchemy==2.0.31
azure-identity==1.17.1